implementa validações de relacionamentos ativos nos métodos de deleção de todas as classes

+++ src/main/java/habilitpro/services/FuncaoService.java
@@ -9,6 +9,7 @@ import org.apache.logging.log4j.Logger;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.NoResultException;
+import javax.persistence.PersistenceException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -51,6 +52,10 @@ public class FuncaoService {
         Funcao funcao = this.funcaoDao.getById(id);
         validaFuncaoNula(funcao);
         this.LOG.info("Função encontrada com sucesso!");
+        if (funcaoDao.verificaSeFuncaoPossuiTrabalhadoresAtivos(id)) {
+            this.LOG.error("A Função não pode ser excluida porque possui Trabalhadores ativos relacionadados!");
+            throw new PersistenceException("The Role cannot be excluded because it has active related Employees!");
+        }
 
         getBeginTransaction();
         this.funcaoDao.delete(funcao);
